requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.5
nltk==3.8.1
packaging==24.0
parsel==1.9.0
pillow==10.2.0
Protego==0.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
pyspellchecker==0.8.1
queuelib==1.6.2
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
requests-file==2.0.0
rsa==4.9
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
textblob==0.18.0.post0
tldextract==5.1.1
tqdm==4.66.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.10.0
urllib3==2.2.1
w3lib==2.1.2
Werkzeug==3.0.1
zope.interface==6.2
